app/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pydantic==1.8.2
elasticsearch==7.15.2
pyyaml==5.4.1
lorem==0.1.1
names==0.3.0
requests==2.25.1